services/training-service/requirements.txt
-e ../../libs/common
fastapi>=0.110
uvicorn[standard]>=0.30
asyncpg>=0.29
structlog>=25.0
httpx>=0.27
opentelemetry-api>=1.25
opentelemetry-sdk>=1.25
opentelemetry-exporter-otlp>=1.25
opentelemetry-instrumentation-fastapi>=0.46b0
opentelemetry-instrumentation-httpx>=0.46b0
nats-py>=2.9.0
pyyaml>=6.0.1
